backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tzdata==2025.1
Werkzeug==3.1.3
gunicorn
requests
python-dotenv==1.0.1
supabase==2.3.4
mysql-connector-python
flask-jwt-extended
flask-bcrypt